medical_imaging_platform/requirements.txt
# Phase 1 - HU-Based Segmentation (ACTIVE)
scipy
numpy
pydicom
trimesh
scikit-image
pyyaml

# Phase 2A - Pre-Trained Models (DISABLED for space; reinstall when ready)
# totalsegmentator>=1.5.0
# torch>=2.0.0
# onnxruntime>=1.14.0

# Phase 2B/2C - Data Collection & Fine-Tuning (DISABLED for space; reinstall when ready)
# monai>=1.0.0
# nnunetv2>=2.0.0
# nibabel>=4.0.0
# SimpleITK>=2.2.0

# Utilities (Core)
tqdm
matplotlib

# TCIA Dataset API (reinstall if using notebook)
# tcia_utils
